Working long polling + echo bot

+++ bot/src/main/kotlin/chonx/bot/telegram/types/Update.kt
@@ -1,4 +1,4 @@
 package chonx.bot.telegram.types
 
 // TODO incomplete
-data class Update(val update_id: Int, val message: Message?)
+data class Update(val update_id: Long, val message: Message?)
